implement named groups of modes
adds GROUP-ALIST to ezk-defkeymaps. Any car of a member on GROUP-ALIST eq to a
symbol in the cdr of a terminal form will have that member's cdr substitute the
symbol as a spliced list

e.g.

(ezk-defkeymaps (<unimplemented-precedence-list>)
 ((CC c-mode c++-mode)
  (LISP emacs-lisp-mode scheme-mode-hook common-lisp-mode))

 ("some keys" ((lambda () (interactive) (message "i'm a lisp")) LISP))
 ("some keys" ((lambda () (interactive) (message "i'm not a lisp")) CC))
 ("some keys" ((lambda () (interactive) (message "i'm something")) LISP CC))
)

equivalent to

(ezk-defkeymaps (<unimplemented-precedence-list>)
 ((CC c-mode c++-mode)
  (LISP emacs-lisp-mode scheme-mode-hook common-lisp-mode))

 ("some keys" ((lambda () (interactive) (message "i'm a lisp"))
        emacs-lisp-mode scheme-mode-hook common-lisp-mode))
 ("some keys" ((lambda () (interactive) (message "i'm not a lisp"))
        c-mode c++-mode))
 ("some keys" ((lambda () (interactive) (message "i'm something"))
        emacs-lisp-mode scheme-mode-hook common-lisp-mode
        c-mode c++-mode)))

This was done because it's common to have a group of similar modes behave
similarly. Named groups are easier to read and write.